Improve method of play script

+++ forger-workflow/src/main/kotlin/com/github/wenslo/forger/workflow/entity/PlayScriptNode.kt
@@ -8,7 +8,7 @@ import com.github.wenslo.forger.workflow.enums.IsFlag
  * @date 2022/10/27 09:05
  */
 data class PlayScriptNode(
-    var playScriptId: Int = 0,
+    var playScriptId: Long = 0,
     var playScriptUniqueId: String = "",
     var uniqueId: String = "",
     var name: String = "",
